+++ it/server/src/test/java/com/walmartlabs/concord/it/server/AnsibleEventIT.java
@@ -78,6 +78,6 @@ public class AnsibleEventIT extends AbstractServerIT {
             return msg.equals("Hi there!");
         }).count();
 
-        assertEquals(1, cnt);
+            assertEquals(1, cnt);
     }
 }