+++ server/api/src/main/java/com/walmartlabs/concord/server/api/process/FormListEntry.java
@@ -9,9 +9,9 @@ package com.walmartlabs.concord.server.api.process;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -26,6 +26,7 @@ import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.io.Serializable;
+import java.util.Map;
 
 @JsonInclude(Include.NON_NULL)
 public class FormListEntry implements Serializable {
@@ -34,17 +35,20 @@ public class FormListEntry implements Serializable {
     private final String name;
     private final boolean custom;
     private final boolean yield;
+    private final Map<String, Object> runAs;
 
     @JsonCreator
     public FormListEntry(@JsonProperty("formInstanceId") String formInstanceId,
                          @JsonProperty("name") String name,
                          @JsonProperty("custom") boolean custom,
-                         @JsonProperty("yield") boolean yield) {
+                         @JsonProperty("yield") boolean yield,
+                         @JsonProperty("runAs") Map<String, Object> runAs) {
 
         this.formInstanceId = formInstanceId;
         this.name = name;
         this.custom = custom;
         this.yield = yield;
+        this.runAs = runAs;
     }
 
     public String getFormInstanceId() {
@@ -63,6 +67,10 @@ public class FormListEntry implements Serializable {
         return yield;
     }
 
+    public Map<String, Object> getRunAs() {
+        return runAs;
+    }
+
     @Override
     public String toString() {
         return "FormListEntry{" +
@@ -70,6 +78,7 @@ public class FormListEntry implements Serializable {
                 ", name='" + name + '\'' +
                 ", custom=" + custom +
                 ", yield=" + yield +
+                ", runAs=" + runAs +
                 '}';
     }
 }